Jetpack Connect:
When we enter the authorization screen for Jetpack connect, let's
only query the site in question, rather than all sites. This will make
for a faster page load.

Additionally, to avoid conflicts when reconnecting, we'll need to update
the site's reducer to respond to a failed single site query by removing
the failed site from the site's list.

After connection, only fetch the site in question.